+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/pki/X509Certificate.kt
@@ -6,7 +6,9 @@ import at.asitplus.signum.indispensable.CryptoSignature
 import at.asitplus.signum.indispensable.X509SignatureAlgorithm
 import at.asitplus.signum.indispensable.asn1.*
 import at.asitplus.signum.indispensable.asn1.encoding.*
+import at.asitplus.signum.indispensable.io.Base64Strict
 import at.asitplus.signum.indispensable.io.ByteArrayBase64Serializer
+import at.asitplus.signum.indispensable.io.TransformingSerializerTemplate
 import at.asitplus.signum.indispensable.pki.AlternativeNames.Companion.findIssuerAltNames
 import at.asitplus.signum.indispensable.pki.AlternativeNames.Companion.findSubjectAltNames
 import at.asitplus.signum.indispensable.pki.TbsCertificate.Companion.Tags.EXTENSIONS
@@ -14,8 +16,10 @@ import at.asitplus.signum.indispensable.pki.TbsCertificate.Companion.Tags.ISSUER
 import at.asitplus.signum.indispensable.pki.TbsCertificate.Companion.Tags.SUBJECT_UID
 import io.matthewnelson.encoding.base64.Base64
 import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
+import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
+import kotlinx.serialization.builtins.serializer
 
 /**
  * Very simple implementation of the meat of an X.509 Certificate:
@@ -33,8 +37,10 @@ constructor(
     val validUntil: Asn1Time,
     val subjectName: List<RelativeDistinguishedName>,
     val publicKey: CryptoPublicKey,
-    val issuerUniqueID: BitSet? = null,
-    val subjectUniqueID: BitSet? = null,
+    @Serializable(with = Asn1TimeSerializer::class)
+    val issuerUniqueID: Asn1BitString? = null,
+    @Serializable(with = Asn1TimeSerializer::class)
+    val subjectUniqueID: Asn1BitString? = null,
     val extensions: List<X509CertificateExtension>? = null,
 ) : Asn1Encodable<Asn1Sequence> {
 
@@ -82,8 +88,8 @@ constructor(
             //subject public key
             +publicKey
 
-            issuerUniqueID?.let { +(Asn1BitString(it) withImplicitTag ISSUER_UID) }
-            subjectUniqueID?.let { +(Asn1BitString(it) withImplicitTag SUBJECT_UID) }
+            issuerUniqueID?.let { +(it withImplicitTag ISSUER_UID) }
+            subjectUniqueID?.let { +(it withImplicitTag SUBJECT_UID) }
 
             extensions?.let {
                 if (it.isNotEmpty()) {
@@ -194,8 +200,8 @@ constructor(
                 validUntil = timestamps.second,
                 subjectName = subject,
                 publicKey = cryptoPublicKey,
-                issuerUniqueID = issuerUniqueID?.toBitSet(),
-                subjectUniqueID = subjectUniqueID?.toBitSet(),
+                issuerUniqueID = issuerUniqueID,
+                subjectUniqueID = subjectUniqueID,
                 extensions = extensions,
             )
         }
@@ -306,3 +312,14 @@ typealias CertificateChain = List<X509Certificate>
 
 val CertificateChain.leaf: X509Certificate get() = first()
 val CertificateChain.root: X509Certificate get() = last()
+
+private
+/** De-/serializes Base64 strings to/from [ByteArray] */
+object Asn1BitStringSerializer : TransformingSerializerTemplate<Asn1BitString?, String>(
+    parent = String.serializer(),
+    encodeAs = { if (it == null) "" else byteArrayOf(it.numPaddingBits, *it.rawBytes).encodeToString(Base64Strict) },
+    decodeAs = {
+        if (it == "") null
+        else Asn1BitString.decodeFromTlv(Asn1Primitive(Asn1Element.Tag.BIT_STRING, it.decodeToByteArray(Base64Strict)))
+    }
+)